+++ CHANGELOG.rst
@@ -45,6 +45,7 @@ Fixed
     - Crash at exception handling in ``clean_new_owner()`` (#981)
     - Incorrect button icon in remote site form (#1001)
     - Case-sensitive project list sorting (#1006)
+    - Project list filtering not trimmed (#1021)
 - **Timeline**
     - Uncaught exceptions in ``get_plugin_lookup()`` (#979)